requirements.txt
Django==4.2.26
djangorestframework==3.16.1
django-cors-headers==4.6.0
python-decouple==3.8
requests==2.32.5
asgiref==3.10.0
